Refactor booking Cypress tests (split drop-out tests)
